Keep semantic tokenizer outside action surface
